requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
matplotlib>=3.4.3
Pillow>=8.3.2
opencv-python>=4.5.3
tqdm>=4.62.3
gradio>=3.1.0